Update both files
In main_model_js.html, there are some tasks  configuring about the javascript functions and elements inside of the html snippet code.
Configuring about the input and output images and not numbers from users, configure and connect the images and also for running the model and not the prediction model elements.
I think I might need to see the other examples code that connects and configures about the code.